pyproject.toml
[project]
name = "protein-engineering-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib==2.208.0",
    "cdk-nag>=2.37.9",
    "constructs>=10.0.0,<11.0.0",
    "mcp>=1.13.0",
    "strands-agents>=1.5.0",
]

[dependency-groups]
dev = [
    "aws-cdk-lib==2.208.0",
    "boto3>=1.40.12",
    "moto[cloudwatch,s3,sagemaker]==4.2.14",
    "pytest==6.2.5",
    "pytest-cov==4.0.0",
    "pytest-html==3.1.1",
    "pytest-mock>=3.14.1",
    "pytest-timeout==2.1.0",
    "pytest-xdist==3.3.1",
]

[tool.uv.workspace]
members = [
    "agent",
]
